requirements/dev.txt
-r base.txt

django-debug-toolbar==0.11.0
-e git+https://github.com/orf/django-debug-toolbar-template-timings.git@2e6d5c29c268cac4b707991e15264b0f915727e9#egg=django_debug_toolbar_template_timings-dev
-e git+https://github.com/playfire/django-debug-toolbar-user-panel.git@29a163dc985c4b173ead57b09b3cfdd6ca84e76f#egg=django_debug_toolbar_user_panel-dev

# fabric integration
fabric
termcolor
GitPython


# Testing
mock
model_mommy
pyquery
django-inmemorystorage